% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{AF16xHK104cross}
\alias{AF16xHK104cross}
\title{Cross object for AF16xHK104 C. brggsae RIAILs}
\format{A cross object made with the \code{qtl} package, with genotype and
phenotype subobjects.}
\description{
In this object, AF16 genotypes are encoded as 1 and HK104 genotypes are
encoded as 2. This cross object contains no heterozyous loci. When the
extract genotype function is run on this cross object, AF16 genotypes are
converted to -1 and HK104 genotypes to 1. For mappings with this cross
object, negative effect sizes indicate that Af16 had the greater phenotype
value while postive effect sizes indicate that HK104 had a greater phenotype
value.
}
