% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmapping.R
\name{checkeQTLintervals}
\alias{checkeQTLintervals}
\title{Find expression QTL from the Rockman dataset within a user-defined interval}
\usage{
checkeQTLintervals(chrom, left_pos, right_pos)
}
\arguments{
\item{chrom}{The chromosome of interest, given as an arabic numeral}

\item{left_pos}{The left position of the interval}

\item{right_pos}{The right position of the interval}
}
\value{
Assigns two data frames to the global environment. eQTL_PositionInfo
contains quantitative information about detected eQTL in the interval.
eQTL_GeneDescriptions contains GO terms and other qualitative data.
}
\description{
Find expression QTL from the Rockman dataset within a user-defined interval
}
