% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{count_strains_per_trait}
\alias{count_strains_per_trait}
\title{Count number of strains with data for each phenotype from cross structure}
\usage{
count_strains_per_trait(pheno)
}
\arguments{
\item{pheno}{The phenotype elemnt of the cross object}
}
\value{
A named vector of the number of strains present for each phenotype
measurement
}
\description{
Count number of strains with data for each phenotype from cross structure
}
