% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{effectplot}
\alias{effectplot}
\title{Plot effect size by marker}
\usage{
effectplot(cross, map, parental = "N2xCB4856")
}
\arguments{
\item{cross}{The cross object used for the mapping}

\item{map}{The mapping output for a single trait}

\item{parental}{The parental strains used to generate RIAILs. Either
"N2xCB4856" (default), "N2xLSJ2", or "AF16xHK104}
}
\value{
A line plot of the effect size at each marker
}
\description{
Plot effect size by marker
}
