% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{extract_genotype}
\alias{extract_genotype}
\title{Extract genotype matrix from cross structure and recode as -1, 1}
\usage{
extract_genotype(cross)
}
\arguments{
\item{cross}{A cross object}
}
\value{
The genotype matrix, encoded as -1 or 1 for genotype
}
\description{
Extract genotype matrix from cross structure and recode as -1, 1
}
