% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{extract_scaled_phenotype}
\alias{extract_scaled_phenotype}
\title{Extract phenotype matrix from cross object, mean center and optionally
standardize variance}
\usage{
extract_scaled_phenotype(cross, set = NULL, setcorrect = FALSE,
  scalevar = TRUE)
}
\arguments{
\item{cross}{A cross object}

\item{set}{A vector of set IDs for all RIAILs}

\item{setcorrect}{Boolean, whether or not to correct for sets by scaling
phenotype values based on the set ID}

\item{scalevar}{Boolean, whether or not to standarize the variance}
}
\value{
A matrix of scaled phenotype values
}
\description{
Extract phenotype matrix from cross object, mean center and optionally
standardize variance
}
