% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmapping.R
\name{findindels}
\alias{findindels}
\title{Find indels to generate primers}
\usage{
findindels(left, right, chr)
}
\arguments{
\item{left}{The left position of the N2 region of interest}

\item{right}{The right position of the N2 region of interest}

\item{chr}{The chromosome of interest, written as a quoted roman numeral}
}
\value{
Outputs a dataframe with all known indel information
}
\description{
Find indels to generate primers
}
