% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmapping.R
\name{findN2fosmids}
\alias{findN2fosmids}
\title{Find N2 fosmids that tile across a given interval}
\usage{
findN2fosmids(chrom, left_pos, right_pos)
}
\arguments{
\item{chrom}{The chromosome of interest, given as a roman numeral value in quotes}

\item{left_pos}{The left position of the interval}

\item{right_pos}{The right position of the interval}
}
\value{
A plot of N2 fosmids that tile across the interval of interest and a data frame
of chromosome, clone name, fosmid start, fosmid end, and where the fosmid can
be found in the Andersen lab.
}
\description{
Find N2 fosmids that tile across a given interval
}
