% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postmapping.R
\name{FindRIAILsforNILs}
\alias{FindRIAILsforNILs}
\title{Find RIAILs for generating NILs to isolate a genomic region}
\usage{
FindRIAILsforNILs(CI.L, CI.R, chromosome)
}
\arguments{
\item{CI.L}{The left marker of the interval, e.g. "UCE1-526"}

\item{CI.R}{The right marker of the interval, e.g. "CE1-108"}

\item{chromosome}{The chromosome of the interval as a quoted roman numeral}
}
\value{
Outputs a list of four elements.
CompleteInterval1 - dataframe of strains with a continous genotype block across user input interval
strains also have a breakpoint in the surrounding region used to generate plot "complete".
BreakInterval - dataframe of strains with breaks in the interval and outside the interval
used to generate plot "complete".
Breaks - ggplot object of all RIAILs identified to have break points in and around the interval
Complete - ggplot object of all RIAILs identified to have a continous genotype stretch across the interval
}
\description{
Find RIAILs for generating NILs to isolate a genomic region
}
