% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{mapformat}
\alias{mapformat}
\title{Get the LOD value for each marker based on the correlation between genotype}
\usage{
mapformat(pheno)
}
\arguments{
\item{pheno}{A phenotype data frame output from the easysorter pipeline}
}
\value{
A fully formatted phenotype data frame ready to be joined to the
cross object
}
\description{
Get the LOD value for each marker based on the correlation between genotype
}
