% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{mergepheno}
\alias{mergepheno}
\title{Merge the cross object and the phenotype data frame using a dplyr left_join}
\usage{
mergepheno(cross, phenotype, set = NULL)
}
\arguments{
\item{cross}{A cross object}

\item{phenotype}{The phenotype data frame with the id numbers for each strain}

\item{set}{Filter the phenotype data to one specific set (Rockman=1,
Andersen=2) before joining to the data frame}
}
\value{
A cross object complete with phenotype information
}
\description{
Incoming phenotype data frame must have the following columns:
\code{condition}
\code{trait}
\code{strain}
\code{phenotype}
}
