% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{N2xCB4856cross}
\alias{N2xCB4856cross}
\title{Cross object for N2xCB4856 RIAILs}
\format{A cross object made with the \code{qtl} package, with genotype and
phenotype subobjects.}
\description{
In this object, N2 genotypes are encoded as 1 and CB4856 genotypes are
encoded as 2. This cross object contains no heterozyous loci. When the
extract genotype function is run on this cross object, N2 genotypes are
converted to -1 and CB4856 genotypes to 1. For mappings with this cross
object, negative effect sizes indicate that N2 had the greater phenotype
value while postive effect sizes indicate that CB4856 had a greater phenotype
value.
}
