% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{N2xLSJ2cross}
\alias{N2xLSJ2cross}
\title{Cross object for N2xLSJ2 RIAILs}
\format{A cross object made with the \code{qtl} package, with genotype and
phenotype subobjects.}
\description{
In this object, N2 genotypes are encoded as 1 and LSJ2 genotypes are
encoded as 3. This cross object had heterozyous loci removed (set to NA). If
heterozygous loci are introduced back into the data, they should be encoded
as 2. When the extract genotype function is run on this cross object, N2
genotypes are converted to -1 and LSJ2 genotypes to 1. If heterozygous loci
are introduced back into the data, they will be extracted as the value 0. For
mappings with this cross object, negative effect sizes indicate that N2 had
the greater phenotype value while postive effect sizes indicate that LSJ2 had
a greater phenotype value.
}
